src/players: improve first trump selection turn.
There were some issues in the method that handles the trump selection
for human player:
if user enter for example that 'nciqsc,', the answer was accepted as a
'no' as the first character is a 'n'.
Improve that reading the three first characters (as to be to the third
or 'nn' would be accepted). For that, use the method
'read_n_carac_and_flush'.
As this method replaces the last character by '\0' if this one is '\n'
and a maximal of N characters have been read, we can just test the
length of the resulting string.
If it has a length strictly greater than 1, the answer was not right.