sr/carte: add the method 'get_value_card'.
Return the value of the card according its rank and if it is trump or
not.
We give tow argument: the card and the trump color. If the color of the
card is equal to 'color_trump" then, we return the value of the rank of
the card in the trump ones.
We can think we could use the field 'is_trump' of the card given in
argument.
But this method will also be used on the trump selecting, by virtual
players. Use the field 'is_trump' would require setting the field before
calling the method 'get_value_card' and setting it again after having
called the method (get_value_card'.
Just better to give directly the trump color.