src/players: adding the card to player sorted.
Sort the card when they are added to players.
Normally, reals players have to wait that a player has taken the trump
card. Else that would mix the game if none player take the card.
But, here, if none players take the card, we just take again the order
of the card at the beginning of the round.