src/gestion_jeu_carte: change cards trump status.
Cards have a boolean indicating if it is a trump card.
So, this boolean has to be updated at each trump change.
Add a method to set it to true for all wards of a specific color:
'set_cards_trump_status'.

Add a method to reset the trumps status of all cards:
'reset_cards_trump_status'.