src/gestion_partie: implement the method 'get_next_trick'.
Implement this method.
So, a trick unfolds like that:
First, we need to get the first card. It becomes the master card.
Then, we need to know if the new card is better than the master card. If
yes, it has to be replaced.
We do that for the four players.