src/players: fix issue trump selecting.
The method 'get_player_cards_value' will take two int pointers and
return the sum of the cards values of the trump card and others cards.
However it should not suppose that these pointers have been well
initialized and so should not add the value found to the selecting card
value eventually already set in these pointers.
It should just return the value and the calling method, so
'does_virtual_player_take_card_first_turn' and
'does_virtual_player_take_card_second_turn' should handle that.

This commit fix that.