src/players: fix issue on virtual trump choice.
There were an issue on virtual player trump selecting.
If a player has king and queen trump cards, then a bonus of 20 points
will be added on its points at the end of the round.
So, these points were added on the points that a user has in its man.
But, they should be added to the player only for the trump cards, not
the others ones !