src/players: create a method to get all cards str.
Create a method to get all cards of a player in a string.
That will let me use the appropriate method to display these data.
Indeed, when a human player indicates if he wants to take the trump card,
we would like the display his cards. But, when this is a virtual player,
we would like to display the cards only on debug.
This methods lets me call logger_trace or directly printf when this nis
necessary.

So remove the method 'display_player_cards' too.