src/players: get next card from virtual player.
Implement the method 'take_card_from_virtual_player'.
The implementation of this method is very basic for now: I get all the
possibles cards that the player could play (there is some rules that the
player has to follow to play a card). Then I return the first card
found.