src/gestion_partie: update trick_t struct.
On trick_t, we need to know, when it will be the tile to count the
points won by each teams, which player has played which card of each
trick.
Indeed, that will let us know if a player has done 'belote and re', that
adds 20 points more for the teams.